Rename useQuery to useSelector
